Staging Release (#17)
* Init Repo

- Add Rails app with Vue as frontend.
- Add self-signed certificates for local use.
- Add Docker to make things portable.

This is a functional boilerplate. No interesting stuff is happening.

* Preliminar favicon and such.

* Simple view for Widget Customization

Add a vuetify form to customize the widget

This form will submit the data to the backend. The backend later will insert or update the script tag. This logic is not implemented yet.

* Amazon S3

Add amazon s3 credentials.

Since this apps goes to Heroku, we need a external storage to save the shop configuration for the script tag.

* Update Widget UI

Add functionality to update the shop widget customization.
Add snackbar to visualize response from backend
Add axios for XMLHttpRequests

* Custom Session Repository

Add autoloadable Shop for Shopify Session Storage

This was necessary since the usual setup is to use the model, Shop, but by doing that, ActiveStorage `has_once_attached` method is not found, since is not loaded before the initializers.

* Widget Customization Backend

Add update action method, placeholderish behavior.

UpdateWidgetCustomization Service is pending.

* Activate Storage

Add migrations for Active Storage.

It will handle automagically S3 storage.

* Widget Customization Fields

Add widget customization fields to Shop model.

* Shopify Basic Embedded App Setup

Add basic logic to handle shopify embedded behavior.

* Shop Creation

Add a before hook to add uuid to the shop
Add a after hook to push script to handle widget customization
Add a after hook to create script tags in shopify shop.

* Update issue templates

* Create CODE_OF_CONDUCT.md

* Create CONTRIBUTING.md

* Update CONTRIBUTING.md

* Specs & Test Coverage

* Add preliminar workflow.